+++ core/src/main/java/io/appulse/epmd/java/core/mapper/deserializer/RequestDeserializer.java
@@ -16,48 +16,65 @@
 
 package io.appulse.epmd.java.core.mapper.deserializer;
 
-import io.appulse.epmd.java.core.mapper.DataSerializable;
 import io.appulse.epmd.java.core.mapper.deserializer.exception.DeserializationException;
+import io.appulse.epmd.java.core.mapper.deserializer.exception.InvalidReceivedMessageLengthException;
 import io.appulse.epmd.java.core.mapper.deserializer.exception.InvalidReceivedMessageTagException;
 import io.appulse.epmd.java.core.model.Tag;
-import io.appulse.epmd.java.core.model.TaggedMessage;
+import io.appulse.epmd.java.core.model.request.Request;
 import io.appulse.utils.Bytes;
 
 import lombok.NonNull;
+import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 
 /**
  *
  * @author Artem Labazin
- * @since 0.0.1
+ * @since 0.4.0
  */
-class DataDeserializer implements Deserializer {
+@Slf4j
+class RequestDeserializer implements Deserializer {
 
   @Override
   public <T> T deserialize (@NonNull Bytes bytes, @NonNull Class<T> type) throws DeserializationException {
+    val length = bytes.getShort();
+    if (length != bytes.remaining()) {
+      val message = String.format("Expected length is %d - %d bytes, but actual length is %d bytes.",
+                                  2, length, bytes.remaining());
+      log.error(message);
+      throw new InvalidReceivedMessageLengthException(message);
+    }
+
     T result;
     try {
       result = type.newInstance();
     } catch (IllegalAccessException | InstantiationException ex) {
+      log.error("Deserialized type instantiation error", ex);
       throw new DeserializationException(ex);
     }
 
-    if (result instanceof TaggedMessage) {
-      val expectedTag = ((TaggedMessage) result).getTag();
-      val tag = Tag.of(bytes.getByte());
-      if (!expectedTag.equals(tag)) {
-        val message = String.format("Expected tag is: %s, but actual tag is: %s",
-                                    expectedTag, tag);
-        throw new InvalidReceivedMessageTagException(message);
-      }
+    if (!(result instanceof Request)) {
+      val message = String.format("Deserializing type '%s' is not an instance of '%s'",
+                                  type.getSimpleName(), Request.class.getSimpleName());
+      log.error(message);
+      throw new DeserializationException(message);
+    }
+    val request = (Request) result;
+
+    val tag = Tag.of(bytes.getByte());
+    if (tag != request.getTag()) {
+      val message = String.format("Expected tag is: %s, but actual tag is: %s",
+                                  request.getTag(), tag);
+      log.error(message);
+      throw new InvalidReceivedMessageTagException(message);
     }
 
-    ((DataSerializable) result).read(bytes);
+    request.read(bytes);
     return result;
   }
 
   @Override
   public boolean isApplicable (@NonNull Class<?> type) {
-    return DataSerializable.class.isAssignableFrom(type);
+    return Request.class.isAssignableFrom(type);
   }
 }