+++ core/src/main/java/io/appulse/epmd/java/core/mapper/serializer/MessageSerializer.java
@@ -16,22 +16,15 @@
 
 package io.appulse.epmd.java.core.mapper.serializer;
 
-import static io.appulse.epmd.java.core.model.Tag.UNDEFINED;
-import static io.appulse.utils.BytesUtils.align;
-import static io.appulse.utils.BytesUtils.asBytes;
 import static java.util.Arrays.asList;
-import static java.util.Optional.ofNullable;
 
-import java.nio.ByteBuffer;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
 
-import io.appulse.epmd.java.core.mapper.Message;
-import io.appulse.epmd.java.core.mapper.exception.MessageAnnotationMissingException;
 import io.appulse.epmd.java.core.mapper.serializer.exception.NoApplicableSerializerException;
 
+import lombok.NonNull;
 import lombok.SneakyThrows;
-import lombok.val;
 
 /**
  *
@@ -44,38 +37,19 @@ public final class MessageSerializer {
 
   static {
     SERIALIZERS = new CopyOnWriteArrayList<>(asList(
+        new RequestSerializer(),
         new DataSerializer(),
         new EnumSerializer()
     ));
   }
 
   @SneakyThrows
-  public byte[] serialize (Object obj) {
-    Class<?> type = ofNullable(obj)
-        .map(Object::getClass)
-        .orElseThrow(NullPointerException::new);
-
-    val annotation = ofNullable(type)
-        .map(it -> it.getAnnotation(Message.class))
-        .orElseThrow(MessageAnnotationMissingException::new);
-
-    val body = SERIALIZERS.stream()
+  public byte[] serialize (@NonNull Object obj) {
+    Class<?> type = obj.getClass();
+    return SERIALIZERS.stream()
         .filter(it -> it.isApplicable(type))
         .findAny()
         .orElseThrow(NoApplicableSerializerException::new)
         .serialize(obj, type);
-
-    val length = annotation.value() == UNDEFINED
-                 ? body.length
-                 : body.length + Byte.BYTES;
-
-    val buffer = ByteBuffer.allocate(annotation.lengthBytes() + length);
-    if (annotation.lengthBytes() > 0) {
-      buffer.put(align(asBytes(length), annotation.lengthBytes()));
-    }
-    if (annotation.value() != UNDEFINED) {
-      buffer.put(annotation.value().getCode());
-    }
-    return buffer.put(body).array();
   }
 }