+++ server/src/main/java/io/appulse/epmd/java/server/command/server/ServerCommandExecutor.java
@@ -16,20 +16,28 @@
 
 package io.appulse.epmd.java.server.command.server;
 
+import static io.netty.channel.ChannelOption.SO_BACKLOG;
+import static io.netty.channel.ChannelOption.SO_KEEPALIVE;
+import static io.netty.channel.ChannelOption.TCP_NODELAY;
 import static java.util.Optional.of;
 import static lombok.AccessLevel.PRIVATE;
 
 import java.io.Closeable;
-import java.io.IOException;
-import java.net.ServerSocket;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 
 import io.appulse.epmd.java.server.cli.CommonOptions;
 import io.appulse.epmd.java.server.command.AbstractCommandExecutor;
 import io.appulse.epmd.java.server.command.CommandOptions;
+import io.appulse.epmd.java.server.command.server.handler.RequestDecoder;
+import io.appulse.epmd.java.server.command.server.handler.ResponseEncoder;
+import io.appulse.epmd.java.server.command.server.handler.ServerHandler;
 
+import io.netty.bootstrap.ServerBootstrap;
+import io.netty.channel.ChannelInitializer;
+import io.netty.channel.EventLoopGroup;
+import io.netty.channel.nio.NioEventLoopGroup;
+import io.netty.channel.socket.SocketChannel;
+import io.netty.channel.socket.nio.NioServerSocketChannel;
 import lombok.NonNull;
 import lombok.SneakyThrows;
 import lombok.experimental.FieldDefaults;
@@ -46,11 +54,11 @@ import lombok.val;
 @FieldDefaults(level = PRIVATE, makeFinal = true)
 public class ServerCommandExecutor extends AbstractCommandExecutor implements Closeable {
 
-  ServerSocket serverSocket;
+  ServerState serverState;
 
-  ExecutorService executor;
+  EventLoopGroup bossGroup;
 
-  Context context;
+  EventLoopGroup workerGroup;
 
   @NonFinal
   volatile boolean closed;
@@ -63,42 +71,45 @@ public class ServerCommandExecutor extends AbstractCommandExecutor implements Cl
         .map(it -> (ServerCommandOptions) it)
         .orElse(new ServerCommandOptions());
 
-    serverSocket = new ServerSocket(getPort());
-
-    executor = Executors.newCachedThreadPool();
-
-    context = Context.builder()
+    serverState = ServerState.builder()
         .nodes(new ConcurrentHashMap<>())
         .commonOptions(commonOptions)
         .serverOptions(serverOptions)
         .build();
+
+    bossGroup = new NioEventLoopGroup(1);
+    workerGroup = new NioEventLoopGroup(2);
   }
 
   @Override
-  @SneakyThrows
   public void execute () {
-    log.debug("Server before start context: {}", context);
+    log.debug("Starting server on port {}", getPort());
     try {
-      while (true) {
-        log.debug("Waiting new connection");
+      new ServerBootstrap()
+          .group(bossGroup, workerGroup)
+          .channel(NioServerSocketChannel.class)
+          .childHandler(new ChannelInitializer<SocketChannel>() {
 
-        val socket = serverSocket.accept();
-        log.debug("New connection was accepted");
-
-        if (context.getAddresses().contains(socket.getInetAddress())) {
-          socket.close();
-          continue;
-        }
-
-        if (isDebug()) {
-          System.out.println();
-        }
-
-        val handler = new ServerWorker(socket, context);
-        executor.execute(handler);
-      }
+            @Override
+            public void initChannel (SocketChannel channel) throws Exception {
+              channel.pipeline()
+                  .addLast("decoder", new RequestDecoder())
+                  .addLast("encoder", new ResponseEncoder())
+                  .addLast("handler", new ServerHandler(serverState));
+            }
+          })
+          .option(SO_BACKLOG, 128)
+          .childOption(SO_KEEPALIVE, true)
+          .childOption(TCP_NODELAY, true)
+          .bind(getPort())
+          .sync()
+          // Wait until the server socket is closed.
+          .channel().closeFuture().sync();
+    } catch (InterruptedException ex) {
+      log.error("Server work exception", ex);
     } finally {
-      close();
+      workerGroup.shutdownGracefully();
+      bossGroup.shutdownGracefully();
     }
   }
 
@@ -106,23 +117,17 @@ public class ServerCommandExecutor extends AbstractCommandExecutor implements Cl
   @SneakyThrows
   public void close () {
     if (closed) {
+      log.debug("Server was already closed");
       return;
     }
     closed = true;
 
-    log.debug("Closing server...");
+    workerGroup.shutdownGracefully();
+    bossGroup.shutdownGracefully();
+    log.debug("Boss and workers groups are closed");
 
-    try {
-      serverSocket.close();
-    } catch (IOException ex) {
-    }
-    log.debug("Server socket was closed");
-
-    if (!executor.isShutdown()) {
-      executor.shutdown();
-    }
-
-    context.getNodes().clear();
+    serverState.getNodes().clear();
+    log.debug("All nodes are removed");
 
     log.info("Server was closed");
   }