+++ server/src/main/java/io/appulse/epmd/java/server/command/server/handler/command/GetNodeInfoRequestHandler.java
@@ -14,26 +14,43 @@
  * limitations under the License.
  */
 
-package io.appulse.epmd.java.server.command.server.handler;
+package io.appulse.epmd.java.server.command.server.handler.command;
 
 import static io.appulse.epmd.java.core.model.Tag.PORT_PLEASE2_REQUEST;
+import static io.netty.channel.ChannelFutureListener.CLOSE;
 
 import io.appulse.epmd.java.core.model.Tag;
 import io.appulse.epmd.java.core.model.request.GetNodeInfo;
+import io.appulse.epmd.java.core.model.request.Request;
 import io.appulse.epmd.java.core.model.response.NodeInfo;
-import io.appulse.epmd.java.server.command.server.Request;
+import io.appulse.epmd.java.server.command.server.ServerState;
 
+import io.netty.channel.ChannelHandlerContext;
 import lombok.NonNull;
+import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 
+/**
+ *
+ * @author Artem Labazin
+ * @since 0.4.0
+ */
+@Slf4j
 class GetNodeInfoRequestHandler implements RequestHandler {
 
   @Override
-  public void handle (@NonNull Request request) {
-    val getNodeInfo = request.parse(GetNodeInfo.class);
+  public void handle (@NonNull Request request, @NonNull ChannelHandlerContext context, @NonNull ServerState state) {
+    log.debug("Processing {}", request);
 
-    val node = request.getContext()
-        .getNodes()
+    if (!(request instanceof GetNodeInfo)) {
+      val message = String.format("Invalid request object:%n%s", request);
+      log.error(message);
+      throw new IllegalArgumentException(message);
+    }
+
+    val getNodeInfo = (GetNodeInfo) request;
+
+    val node = state.getNodes()
         .get(getNodeInfo.getName());
 
     NodeInfo response;
@@ -52,8 +69,10 @@ class GetNodeInfoRequestHandler implements RequestHandler {
           .name(node.getName())
           .build();
     }
+    log.debug("Response: {}", response);
 
-    request.respondAndClose(response);
+    context.writeAndFlush(response)
+        .addListener(CLOSE);
   }
 
   @Override